requirements.txt
claude-agent-sdk
python-dotenv
fastapi
uvicorn[standard]
pydantic
anyio
langsmith
openinference-instrumentation
